Initial implementation of proposals
This commit adds an implementation of proposal_create_operation and
declares two new operations, proposal_update_operation and
proposal_delete_operation, which are not yet implemented.

It also adds a method to all operations which finds the root authorities
whose approvals are necessary and sufficient for that operation to
execute, as well as a visitor to find the set of root authorities for a
group of operations. This will be valuable when evaluating whether a
proposed transaction has accrued sufficient approval to execute or not.